Separate Cairo0 execution
